RouteController test refactoring (#390)
Added sort by expectedCompletion for RoutePoints in Route

+++ goods-partner-backend/src/main/java/com/goodspartner/web/controller/RouteController.java
@@ -49,7 +49,7 @@ public class RouteController {
 
     @PreAuthorize("hasAnyRole('ADMIN', 'LOGISTICIAN', 'DRIVER')")
     @PostMapping("/{id}/{action}")
-    public RouteActionResponse apply(@PathVariable int id,
+    public RouteActionResponse apply(@PathVariable long id,
                                      @PathVariable String action) {
         return routeService.updateRoute(id, RouteAction.of(action));
     }
